New hypervisor implementation: chroot manager
This patch adds a new hypervisor implementation: a chroot manager. This
hypervisor type can be used to manage (in combination with special OS
definitions) the start and stop of chroot areas, and if used with drbd
instances, it allows (via failover) the migration of chroots between
nodes.

This is a work in progress, and the way chroots should work is not very
clear and does not fit very well in the OS definition framework.
However, the hypervisor works and (if the sshd in the chroot is well
configured) it allows login to the instance both via ssh and console as
for a normal instance.

TODOs:
  - implement instance IP add/remove to/from the bridge, if the instance
    has a defined IP
  - investigate improvements to the OS API so that the create script has
    more information available, e.g. about the hypervisor type
  - mount extra disks in the chroot or alternatively refuse to start
    with more than one disk

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Guido Trotter <ultrotter@google.com>